Extract type origin analysis
