+++ crates/masm-decompiler/src/types/mod.rs
@@ -6,6 +6,7 @@ mod declared;
 mod domain;
 mod inter;
 mod intra;
+mod origin;
 mod stdlib;
 mod summary;
